refactor: 페이지 내부 섹션 구분
