+++ src/refactoring/pages/admin/components/index.tsx
@@ -1,4 +1,4 @@
-export * from "./add-new.product";
+export * from "./product-management/add-new.product";
 export * from "./coupon-management";
-export * from "./product-accordion";
+export * from "./product-management/product-accordion";
 export * from "./product-management";
