+++ src/refactoring/pages/admin/components/product-management/add-new.product.tsx
