+++ src/refactoring/pages/admin/components/product-management/index.tsx
@@ -1,6 +1,6 @@
 import { useProductContext } from "@r/model/product/product-context";
-import { ProductAccordion } from "./product-accordion";
 import { AddNewProduct } from "./add-new.product";
+import { ProductAccordion } from "./product-accordion";
 
 interface ProductManagementProps {}
