+++ src/refactoring/pages/admin/components/product-management/product-accordion.tsx
