+++ src/refactoring/pages/cart/components/index.tsx
@@ -1,5 +1,4 @@
 export * from "./cart-list";
-export * from "./coupon-applier";
-export * from "./order-summary";
-export * from "./product-list";
-export * from "./product-card";
+export * from "./cart-list/coupon-applier";
+export * from "./cart-list/order-summary";
+export * from "./product-list/product-card";
